Addressed possible issue that might occur
if you attempt to call the local writer in the object and pass some like this

$this.WriteHostWriter("something {0} {1}" -f $v, $b)

This would fail as it sees that we are trying to pass two things when we can only pass one. None of the things i addressed cause this, but incase it changes later on down the road, it wouldn't occur.