Updated Confirm-ExchangeShell to handle Errors
Added the CatchActionFunction to be able to call if we have a custom action that we would like to take for if a catch occurs to account for errors in a script.

Because of this, we also want to account for the errors that could occur in the RemoteExchange.ps1 script.